requirements.txt
# Core dependencies
python-dotenv==1.0.0
rich==13.7.0

# LLM providers
google-generativeai==0.8.0
requests==2.31.0

# HTTP client
httpx==0.27.0

# Data validation
pydantic==2.5.0
